Move printer into own file
