+++ src/Language/Mimsa.hs
@@ -7,12 +7,11 @@ module Language.Mimsa
     mkName,
     StringType (..),
     parseExpr,
-    prettyPrint,
     repl,
   )
 where
 
+import Language.Mimsa.Parser
 import Language.Mimsa.Repl (repl)
-import Language.Mimsa.Syntax
 import Language.Mimsa.Typechecker (doInference, startInference)
 import Language.Mimsa.Types
