+++ src/Language/Mimsa/Actions.hs
@@ -13,9 +13,9 @@ import Control.Monad (join)
 import Data.Bifunctor (first)
 import qualified Data.Map as M
 import Data.Text (Text)
+import Language.Mimsa.Parser (parseExpr)
 import Language.Mimsa.Project (getCurrentBindings)
 import Language.Mimsa.Store (createStoreExpression, substitute)
-import Language.Mimsa.Syntax (parseExpr)
 import Language.Mimsa.Typechecker
 import Language.Mimsa.Types
