+++ src/Language/Mimsa/Library.hs
@@ -11,6 +11,7 @@ where
 import Data.Coerce
 import qualified Data.Map as M
 import qualified Data.Text.IO as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types
 import System.Random
