+++ src/Language/Mimsa/Logging.hs
@@ -2,7 +2,7 @@ module Language.Mimsa.Logging where
 
 import qualified Data.Text as T
 import Debug.Trace
-import Language.Mimsa.Types.Printer
+import Language.Mimsa.Printer
 
 debugLog :: (Show b) => String -> b -> b
 debugLog title item = snd (traceShowId (title, item))
