+++ src/Language/Mimsa/Parser.hs
@@ -1,0 +1,8 @@
+module Language.Mimsa.Parser
+  ( module Language.Mimsa.Parser.Language,
+    module Language.Mimsa.Parser.Parser,
+  )
+where
+
+import Language.Mimsa.Parser.Language
+import Language.Mimsa.Parser.Parser
