+++ src/Language/Mimsa/Parser/Language.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Mimsa.Syntax.Language
+module Language.Mimsa.Parser.Language
   ( parseExpr,
     parseExpr',
     expressionParser,
@@ -19,8 +19,8 @@ import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Data.Text as T
-import Language.Mimsa.Syntax.Parser (Parser)
-import qualified Language.Mimsa.Syntax.Parser as P
+import Language.Mimsa.Parser.Parser (Parser)
+import qualified Language.Mimsa.Parser.Parser as P
 import Language.Mimsa.Types
 
 type ParserExpr = Expr Name
