--- src/Language/Mimsa/Syntax.hs
@@ -1,8 +1,0 @@
-module Language.Mimsa.Syntax
-  ( module Language.Mimsa.Syntax.Language,
-    module Language.Mimsa.Syntax.Parser,
-  )
-where
-
-import Language.Mimsa.Syntax.Language
-import Language.Mimsa.Syntax.Parser
