+++ src/Language/Mimsa/Types/Bindings.hs
@@ -8,9 +8,9 @@ import qualified Data.Aeson as JSON
 import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Name
-import Language.Mimsa.Types.Printer
 
 -- a list of names to hashes
 newtype Bindings = Bindings {getBindings :: Map Name ExprHash}
