+++ src/Language/Mimsa/Types/Environment.hs
@@ -5,9 +5,9 @@ module Language.Mimsa.Types.Environment where
 import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Construct
-import Language.Mimsa.Types.Printer
 import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Variable
