+++ src/Language/Mimsa/Types/ExprHash.hs
@@ -5,7 +5,7 @@ module Language.Mimsa.Types.ExprHash where
 
 import qualified Data.Aeson as JSON
 import qualified Data.Text as T
-import Language.Mimsa.Types.Printer
+import Language.Mimsa.Printer
 
 ------------
