+++ src/Language/Mimsa/Types/InterpreterError.hs
@@ -5,9 +5,9 @@ module Language.Mimsa.Types.InterpreterError where
 import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Name
-import Language.Mimsa.Types.Printer
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Variable
