+++ src/Language/Mimsa/Types/ResolverError.hs
@@ -4,9 +4,9 @@ module Language.Mimsa.Types.ResolverError where
 
 import qualified Data.Map as M
 import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Name
-import Language.Mimsa.Types.Printer
 
 data ResolverError = MissingBinding Name Bindings
   deriving (Eq, Ord, Show)
