+++ src/Language/Mimsa/Types/Substitutions.hs
@@ -6,8 +6,8 @@ import Data.Map (Map)
 import qualified Data.Map as M
 import Data.Maybe (fromMaybe)
 import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Printer
 import Language.Mimsa.Types.Variable
 
 ---
