+++ src/Language/Mimsa/Types/TypeName.hs
@@ -7,9 +7,9 @@ module Language.Mimsa.Types.TypeName where
 import qualified Data.Aeson as JSON
 import qualified Data.Text as T
 import GHC.Generics
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.Construct
 import Language.Mimsa.Types.Name
-import Language.Mimsa.Types.Printer
 
 data TypeName = ConsName Construct [TypeName] | VarName Name
   deriving (Eq, Ord, Show, Generic, JSON.ToJSON, JSON.FromJSON)
